+++ src/app/api/componyInfo.ts
@@ -1,6 +1,7 @@
+import { CompanyInfoType } from '@/types/frame';
 import { createClient } from '@/utils/supabase/server';
 
-export const getCompanyInfo = async () => {
+export const getCompanyInfo = async (): Promise<CompanyInfoType[]> => {
   const supabase = await createClient();
   const { data, error } = await supabase.from('company_info').select('*');
   if (error) {
